fix table schema definition

+++ src/database/migrations/2023_06_18_175014_alter_table_users_add_role_id_field.php
@@ -14,7 +14,7 @@ class AlterTableUsersAddRoleIdField extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedBigInteger('role_id')->after('password')->default(1);
+            $table->smallInteger('role_id')->after('password')->default(1);
             $table->foreign('role_id')->references('id')->on('roles');
         });
     }
