+++ src/database/migrations/2024_06_24_063215_create_transaction_details_table.php
@@ -9,10 +9,10 @@ class CreateTransactionDetailsTable extends Migration
     public function up()
     {
         Schema::create('transaction_details', function (Blueprint $table) {
-            $table->unsignedBigInteger('transaction_id');
+            $table->smallInteger('transaction_id');
             $table->mediumInteger('product_id');
-            $table->unsignedBigInteger('unit');
-            $table->integer('qty');
+            $table->smallInteger('unit');
+            $table->smallInteger('qty');
 
             // Foreign key constraints
             $table->foreign('transaction_id')->references('id')->on('transactions');
